refactor(API): use REST-style API routes
BREAKING CHANGE: old routes no more
