+++ apps/web/package.json
@@ -6,7 +6,8 @@
     "dev": "next dev --turbo",
     "build": "tsx lib/db/migrate && next build",
     "start": "next start",
-    "lint": "pnpm lint:next && pnpm lint:biome",
+    "lint": "pnpm lint:types && pnpm lint:next && pnpm lint:biome",
+    "lint:types": "tsc --noEmit",
     "lint:next": "next lint",
     "lint:biome": "biome lint --unsafe",
     "lint:fix": "pnpm lint:next:fix && pnpm lint:biome:fix",
